+++ R/spatSample.r
@@ -10,7 +10,7 @@
 #' 
 #' @param values Logical: If `TRUE` (default), values of the `GRaster` at points are returned.
 #' 
-#' @param cats Logical: If `TRUE` and the `GRaster` is [categorical][tutorial_raster_data_types], then return the category label of each cell. If `values` is also `TRUE`, then the cell value will also be returned.
+#' @param cats Logical: If `TRUE` (default) and the `GRaster` is [categorical][tutorial_raster_data_types], then return the category label of each cell. If `values` is also `TRUE`, then the cell value will also be returned.
 #' 
 #' @param xy Logical: If `TRUE`, return the longitude and latitude of each point. Default is `FALSE`.
 #'
@@ -41,7 +41,7 @@ methods::setMethod(
 		size,
 		as.points = FALSE,
 		values = TRUE,
-		cats = FALSE,
+		cats = TRUE,
 		xy = FALSE,
 		strata = NULL,
 		byStratum = FALSE,
@@ -238,7 +238,7 @@ methods::setMethod(
 	# extract values from raster
 	if (values | cats) {
 
-		vals <- .extractFromRasterAtPoints(x = x, y = src, cats = cats, verbose = verbose)
+		vals <- .extractFromRasterAtPoints(x = x, y = src, values = values, cats = cats, verbose = verbose)
 
 		if (exists("out", inherits = FALSE)) {
 			out <- cbind(out, vals)